Extraer geolocalización de mark.js
Séptimo paso de la descomposición de mark.js: banner GPS inline,
mini-mapa Leaflet, contador regresivo, y los dos flujos de solicitud
(requestGPSBackground al cargar/tras error de paso 2,
requestGPSManual como reintento explícito) se mueven a mark/gps.js.

El módulo no conoce `state` (vive en mark.js, lo usan muchas otras
partes del archivo) — initGps() recibe onLocation (mark.js hace
state.location = coords) y checkEnableMark como callbacks, cableados
una sola vez junto con los 2 botones de reintento GPS (mini-mapa y
banner inline), reemplazando el bloque que antes vivía en EVENT
LISTENERS.

De paso, se deduplica el bloque "resetear ubicación y mapa" que
aparecía idéntico 3 veces en mark.js (returnToSplash, transitionToStep1,
resetSystem) en una sola función exportada, resetLocationUi().

El import de leaflet se mueve de mark.js a gps.js (único lugar que
lo sigue usando). Refactor puro, sin cambios de comportamiento.
